+++ include/gtsam_ext/factors/integrated_color_consistency_factor.hpp
@@ -114,7 +114,7 @@ private:
   double correspondence_update_tolerance_rot;
   double correspondence_update_tolerance_trans;
   mutable Eigen::Isometry3d last_correspondence_point;
-  mutable std::vector<int> correspondences;
+  mutable std::vector<long> correspondences;
 };
 
 using IntegratedColorConsistencyFactor = IntegratedColorConsistencyFactor_<>;
