+++ include/gtsam_ext/factors/integrated_icp_factor.hpp
@@ -116,7 +116,7 @@ private:
   double correspondence_update_tolerance_rot;
   double correspondence_update_tolerance_trans;
   mutable Eigen::Isometry3d last_correspondence_point;
-  mutable std::vector<int> correspondences;
+  mutable std::vector<long> correspondences;
 
   std::shared_ptr<const TargetFrame> target;
   std::shared_ptr<const SourceFrame> source;
