+++ include/gtsam_ext/factors/integrated_loam_factor.hpp
@@ -133,7 +133,7 @@ private:
   double correspondence_update_tolerance_rot;
   double correspondence_update_tolerance_trans;
   mutable Eigen::Isometry3d last_correspondence_point;
-  mutable std::vector<std::tuple<int, int, int>> correspondences;
+  mutable std::vector<std::tuple<long, long, long>> correspondences;
 
   std::shared_ptr<const TargetFrame> target;
   std::shared_ptr<const SourceFrame> source;
@@ -191,7 +191,7 @@ private:
   double correspondence_update_tolerance_rot;
   double correspondence_update_tolerance_trans;
   mutable Eigen::Isometry3d last_correspondence_point;
-  mutable std::vector<std::tuple<int, int>> correspondences;
+  mutable std::vector<std::tuple<long, long>> correspondences;
 
   std::shared_ptr<const TargetFrame> target;
   std::shared_ptr<const SourceFrame> source;
